moved categories factories to their api file
