+++ firstvoices/backend/tests/factories.py
@@ -4,7 +4,6 @@ from django.contrib.auth.models import AnonymousUser
 from factory.django import DjangoModelFactory
 
 from backend.models.app import AppJson, AppMembership
-from backend.models.category import Category
 from backend.models.characters import (
     Alphabet,
     Character,
@@ -154,25 +153,6 @@ class AlphabetFactory(DjangoModelFactory):
 
     class Meta:
         model = Alphabet
-
-
-class ParentCategoryFactory(DjangoModelFactory):
-    site = factory.SubFactory(SiteFactory)
-    site = factory.SubFactory(SiteFactory)
-    title = factory.Sequence(lambda n: "Category title %03d" % n)
-    description = factory.Sequence(lambda n: "Category description %03d" % n)
-    created_by = factory.SubFactory(UserFactory)
-    last_modified_by = factory.SubFactory(UserFactory)
-
-    class Meta:
-        model = Category
-
-
-class ChildCategoryFactory(ParentCategoryFactory):
-    parent = factory.SubFactory(ParentCategoryFactory)
-
-    class Meta:
-        model = Category
 
 
 def get_anonymous_user():
